+++ usr/SAR/results.R
@@ -1,10 +1,10 @@
 # setup -------------------------------------------------------------------
-
+library(dplyr)
 
 # results -----------------------------------------------------------------
-source('scripts/input.R', encoding = 'UTF-8')
-source('scripts/describe.R', encoding = 'UTF-8')
-source('scripts/modeling.R', encoding = 'UTF-8')
-source('scripts/inference.R', encoding = 'UTF-8')
-source('scripts/plots.R', encoding = 'UTF-8')
-# source('scripts/plots-save.R', encoding = 'UTF-8') ## only manual!
+source('scripts/input.R', encoding = 'UTF-8') %>% suppressMessages()
+source('scripts/describe.R', encoding = 'UTF-8') %>% suppressMessages()
+source('scripts/modeling.R', encoding = 'UTF-8') %>% suppressMessages()
+source('scripts/inference.R', encoding = 'UTF-8') %>% suppressMessages()
+source('scripts/plots.R', encoding = 'UTF-8') %>% suppressMessages()
+# source('scripts/plots-save.R', encoding = 'UTF-8') %>% suppressMessages() ## only manual!