+++ NEWS.md
@@ -1,8 +1,9 @@
 # philsf.workflow (development version)
 
 * SAR-links now filters markdown files in report/ #108
-* load philsfmisc in input #110
+* load `philsfmisc` in `input` #110
 * Use current date as default #111
+* Silence `results` script #118
 
 # philsf.workflow 0.11.2